+++ examination-system-common/src/main/java/com/examination_system/model/dto/common/QuestionOptionDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class QuestionOptionDTO {
-    private Integer optionId; // có thể null khi tạo mới
+    private Long optionId; // có thể null khi tạo mới
     private String optionContent;
     @JsonProperty("isCorrect")
     private boolean isCorrect;
