+++ examination-system-common/src/main/java/com/examination_system/model/dto/request/exam/StudentChoiceRequest.java
@@ -16,5 +16,5 @@ public class StudentChoiceRequest {
     @NotNull(message = "isRemove required !")
     boolean isRemove;
     @NotNull
-    Map<Integer, Set<Integer>> studentChoice;
+    Map<Long, Set<Long>> studentChoice;
 }
