+++ examination-system-common/src/main/java/com/examination_system/repository/log/ExamLogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.examination_system.model.entity.log.ExamLog;
 
 @Repository
-public interface ExamLogRepository extends JpaRepository<ExamLog, Integer> {
+public interface ExamLogRepository extends JpaRepository<ExamLog, Long> {
 
 }
